modifies getUniqueFolder to be able to hande paths with ~
The behavior of getUniqueFolder is slightly differnt now, it return the
full path now (with leading "/") and is able to handle shortened $HOME
(~) paths. It still is not able to handle relative paths and "." as
input is ignored.
Theses changes make it possible to make expand the functionality of
`truncate_to_unique` to include `truncate_to_unique_but_last` by adjusting
`POWERLEVEL9K_SHORTEN_DIR_LENGTH` while making the code more compact.
`truncate_to_unique_but_last` is thus obsolete and is being removed.
I also added some docs for getUniqueFolder and propose to move it to the
utilities.zsh where the other truncate functions are as well.

I also added a lot more tests for making sure everything works as expected.